category, price sorting

+++ application.py
@@ -32,7 +32,8 @@ def view_home():
         data = DB.get_items_bycategory(category)
 
     if sort_by == "price":
-        data = {k: v for k, v in sorted(data.items(), key=lambda item: int(item[1]['price']))}
+        data = {k:v for k, v in sorted(data.items(), key=lambda item: float(item[1]['price']))}
+
     item_counts = len(data)
     if item_counts <= per_page:
         data = dict(list(data.items())[:item_counts])
